refactor(analytics): continue decomposing CodebaseAnalytics.vue (#1579)
- Extract AnalyticsHeaderControls.vue (replaces CodebaseAnalyticsHeader)
- Extract AnalyticsProgressSection.vue (replaces CodebaseProgressPanel)
- Extract useIndexingJob.ts composable for indexing state management
- Extract useDashboardLoaders.ts composable for data loading
- Update ProblemsReportSection with export emit support
- Remove orphaned #1469 panel components superseded by this PR
- Reduces CodebaseAnalytics.vue from ~5500 to ~4148 lines (~25%)
